[fix] Answer Page

+++ IQTest_backend/db/migrate/20210628040911_add_marks_to_users.rb
@@ -1,5 +1,5 @@
 class AddMarksToUsers < ActiveRecord::Migration[6.1]
   def change
-    add_column :users, :marks, :decimal
+    add_column :users, :marks, :double(5,2)
   end
 end
